Fixed behaviour near boundaries
Previously the scale was only being calculated once at p instead of four
times at p+dy, p-dy, p+dx, and p-dx. The result was that the gradient of
the falloff near boundaries was not having the appropriate effect.

Additionally, the distance to the boundary was left as a squared value
instead of being rooted which had a significant effect.